Added type hinting, fixed 'copy' issue

+++ docker/container.py
@@ -6,10 +6,8 @@
 # SPDX-License-Identifier: BSD-3-Clause
 
 import argparse
-import os
 import shutil
 import subprocess
-import sys
 from pathlib import Path
 
 from isaaclab_container_utils import apptainer_utils, x11_utils
@@ -23,7 +21,7 @@ def main():
 
     # We have to create a separate parent parser for common options to our subparsers
     parent_parser = argparse.ArgumentParser(add_help=False)
-    parent_parser.add_argument("profile", nargs="?", help="Optional container profile specification.")
+    parent_parser.add_argument("profile", nargs="?", default="base", help="Optional container profile specification.")
 
     subparsers.add_parser(
         "start", help="Build the docker image and create the container in detached mode.", parents=[parent_parser]
@@ -45,97 +43,47 @@ def main():
     args = parser.parse_args()
 
     if not shutil.which("docker"):
-        print("[Error] Docker is not installed! Please check the 'Docker Guide' for instruction.", file=sys.stderr)
-        sys.exit(1)
+        raise RuntimeError("Docker is not installed! Please check the 'Docker Guide' for instruction.")
 
-    base_dir = Path(__file__).resolve().parent
+    context_dir = Path(__file__).resolve().parent
     # Creating statefile
-    statefile = Statefile(statefile=base_dir / ".container.yaml")
-    container_interface = IsaacLabContainerInterface(base_dir=base_dir, profile=args.profile, statefile=statefile)
+    statefile = Statefile(statefile=context_dir / ".container.yaml")
+    container_interface = IsaacLabContainerInterface(context_dir=context_dir, profile=args.profile, statefile=statefile)
 
     print(f"[INFO] Using container profile: {container_interface.profile}")
-
     if args.command == "start":
         print(
             f"[INFO] Building the docker image and starting the container {container_interface.container_name} in the"
             " background..."
         )
-        os.chdir(container_interface.base_dir)
-        container_interface.add_yamls += x11_utils.x11_check(statefile)
-        subprocess.run(
-            [
-                "docker",
-                "compose",
-                "--file",
-                "docker-compose.yaml",
-                "--env-file",
-                ".env.base",
-                "build",
-                "isaac-lab-base",
-            ],
-            check=True,
-        )
-        subprocess.run(
-            ["docker", "compose"]
-            + container_interface.add_yamls
-            + container_interface.add_profiles
-            + container_interface.add_envs
-            + ["up", "--detach", "--build", "--remove-orphans"],
-            check=True,
-        )
+        x11_yaml, x11_envar = x11_utils.x11_check(container_interface.statefile)
+        container_interface.add_yamls += x11_yaml
+        container_interface.environ.update(x11_envar)
+        container_interface.start()
     elif args.command == "enter":
-        container_interface.is_container_running()
         print(f"[INFO] Entering the existing {container_interface.container_name} container in a bash session...")
-        os.chdir(container_interface.base_dir)
-        subprocess.run(
-            ["docker", "exec", "--interactive", "--tty", f"{container_interface.container_name}", "bash"], check=True
-        )
+        container_interface.enter()
     elif args.command == "copy":
-        container_interface.is_container_running()
-        print(f"[INFO] Copying artifacts from the 'isaac-lab-{container_interface.profile}' container...")
-        artifacts = {"logs": "logs", "docs/_build": "docs/_build", "data_storage": "data_storage"}
-        for container_path, host_path in artifacts.items():
-            print(f"\t - /workspace/isaaclab/{container_path} -> {container_interface.base_dir}/artifacts/{host_path}")
-        os.chdir(container_interface.base_dir)
-        for path in artifacts.values():
-            shutil.rmtree(container_interface.base_dir / f"artifacts/{path}", ignore_errors=True)
-        (container_interface.base_dir / "artifacts/docs").mkdir(parents=True, exist_ok=True)
-        for container_path, host_path in artifacts.items():
-            subprocess.run(
-                [
-                    "docker",
-                    "cp",
-                    f"isaac-lab-{container_interface.profile}:/workspace/isaaclab/{container_path}",
-                    f"./artifacts/{host_path}",
-                ],
-                check=True,
-            )
+        print(f"[INFO] Copying artifacts from the 'isaac-lab-{container_interface.container_name}' container...")
+        container_interface.copy()
         print("\n[INFO] Finished copying the artifacts from the container.")
     elif args.command == "stop":
-        container_interface.is_container_running()
         print(f"[INFO] Stopping the launched docker container {container_interface.container_name}...")
-        os.chdir(container_interface.base_dir)
-        subprocess.run(
-            ["docker", "compose", "--file", "docker-compose.yaml"]
-            + container_interface.add_profiles
-            + container_interface.add_envs
-            + ["down"],
-            check=True,
-        )
+        container_interface.stop()
         x11_utils.x11_cleanup(statefile)
     elif args.command == "push":
         if not shutil.which("apptainer"):
             apptainer_utils.install_apptainer()
-        container_interface.check_image_exists()
+        if not container_interface.does_image_exist():
+            raise RuntimeError(f"The image '{container_interface.image_name}' does not exist!")
         apptainer_utils.check_docker_version_compatible()
         cluster_login = container_interface.dot_vars["CLUSTER_LOGIN"]
         cluster_isaaclab_dir = container_interface.dot_vars["CLUSTER_ISAACLAB_DIR"]
         cluster_sif_path = container_interface.dot_vars["CLUSTER_SIF_PATH"]
-        exports_dir = container_interface.base_dir / "exports"
+        exports_dir = container_interface.context_dir / "exports"
         exports_dir.mkdir(parents=True, exist_ok=True)
         for file in exports_dir.glob(f"{container_interface.container_name}*"):
             file.unlink()
-        os.chdir(exports_dir)
         subprocess.run(
             [
                 "APPTAINER_NOHTTPS=1",
@@ -148,12 +96,14 @@ def main():
             ],
             check=True,
             shell=True,
+            cwd=exports_dir,
         )
         subprocess.run(
             ["tar", "-cvf", f"{container_interface.container_name}.tar", f"{container_interface.container_name}.sif"],
             check=True,
+            cwd=exports_dir,
         )
-        subprocess.run(["ssh", cluster_login, f"mkdir -p {cluster_sif_path}"], check=True)
+        subprocess.run(["ssh", cluster_login, f"mkdir -p {cluster_sif_path}"], check=True, cwd=exports_dir)
         subprocess.run(
             [
                 "scp",
@@ -161,6 +111,7 @@ def main():
                 f"{cluster_login}:{cluster_sif_path}/{container_interface.container_name}.tar",
             ],
             check=True,
+            cwd=exports_dir,
         )
     elif args.command == "job":
         cluster_login = container_interface.dot_vars["CLUSTER_LOGIN"]
@@ -175,7 +126,7 @@ def main():
                 "--exclude",
                 "*.git*",
                 "--filter=:- .dockerignore",
-                f"/{container_interface.base_dir}/..",
+                f"/{container_interface.context_dir}/..",
                 f"{cluster_login}:{cluster_isaaclab_dir}",
             ],
             check=True,
@@ -193,8 +144,7 @@ def main():
             check=True,
         )
     else:
-        print(f"[Error] Invalid command provided: {args.command}", file=sys.stderr)
-        sys.exit(1)
+        raise RuntimeError(f"Invalid command provided: {args.command}")
 
 
 if __name__ == "__main__":
